[OBMIN-320] fixed input file
