+++ src/components/error-display/index.tsx
@@ -4,4 +4,5 @@ import * as Styles from './styles';
 const ErrorDisplay = ({ error }: IError) => (
   <> {error && <Styles.StyledError>{error}</Styles.StyledError>}</>
 );
+
 export { ErrorDisplay };
