+++ src/components/input-file/types.ts
@@ -1,8 +1,11 @@
-type InputOnChangeEventType =
+export type InputOnChangeEventType =
   | React.ChangeEvent<HTMLInputElement>
   | React.DragEvent<HTMLLabelElement>;
 
 export interface IFileInput {
+  type: string;
+  name: string;
+  error?: string;
   onChange: (e: InputOnChangeEventType, files?: File[]) => void;
 }
