--- src/components/photo-files/config.js
@@ -1,24 +1,0 @@
-export const MAX_SIZE_PHOTO = 10;
-
-const extension = ['jpg', 'jpeg', 'png', 'gif'];
-
-export const convertToMB = (bytes) => {
-  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB'];
-  if (!bytes) return '0 bytes';
-  const i = parseInt(Math.floor(Math.log(bytes) / Math.log(1024)), 10);
-  return {
-    value: +(bytes / 1024 ** i).toFixed(2),
-    valueString: `${(bytes / 1024 ** i).toFixed(2)} ${sizes[i]}`,
-  };
-};
-
-export const options = {
-  maxSizeMB: MAX_SIZE_PHOTO,
-  maxWidthOrHeight: 1920,
-  useWebWorker: true,
-  maxIteration: 10,
-};
-
-export const isRightExtension = (type) => {
-  return !extension.includes(type.replace('image/', ''));
-};
